+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/implementation/servicebus/properties/AzureServiceBusCommonProperties.java
@@ -13,7 +13,7 @@ import com.azure.spring.cloud.service.servicebus.properties.ServiceBusEntityType
 public abstract class AzureServiceBusCommonProperties extends AbstractAzureAmqpConfigurationProperties {
 
     // https://help.boomi.com/bundle/connectors/page/r-atm-Microsoft_Azure_Service_Bus_connection.html
-    // https://docs.microsoft.com/en-us/rest/api/servicebus/addressing-and-protocol
+    // https://docs.microsoft.com/rest/api/servicebus/addressing-and-protocol
     /**
      * The domain name of a Service Bus namespace.
      */