+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/implementation/graph/UserPrincipalMicrosoftGraphTests.java
@@ -95,7 +95,7 @@ class UserPrincipalMicrosoftGraphTests {
     @Test
     void getGroups() throws Exception {
         properties.getUserGroup().setAllowedGroupNames(Arrays.asList("group1", "group2", "group3"));
-        AzureADGraphClient graphClientMock = new AzureADGraphClient(clientId, clientSecret, properties,
+        AadGraphClient graphClientMock = new AadGraphClient(clientId, clientSecret, properties,
             endpoints);
         stubFor(get(urlEqualTo("/v1.0/me/memberOf"))
             .withHeader(ACCEPT, equalTo(APPLICATION_JSON_VALUE))